+++ src/ManagedModels/States/Archivable/Archivable.php
@@ -8,29 +8,34 @@ use Thinktomorrow\Chief\ManagedModels\States\PageState\PageState;
 
 trait Archivable
 {
-    protected function getArchivablePageStateAttribute(): string
+    protected static function getArchiveColumn(): string
     {
         return PageState::KEY;
     }
 
+    protected static function getArchivedValue(): string
+    {
+        return PageState::archived->value;
+    }
+
     protected static function bootArchivable()
     {
-        static::addGlobalScope(new ArchiveScope);
+        static::addGlobalScope(new ArchiveScope(static::getArchiveColumn(), static::getArchivedValue()));
     }
 
     public function isArchived(): bool
     {
-        return $this->getState(\Thinktomorrow\Chief\ManagedModels\States\PageState\PageState::KEY) === PageState::archived;
+        return $this->getState(static::getArchiveColumn())->value === static::getArchivedValue();
     }
 
     public function scopeArchived($query)
     {
-        $query->withoutGlobalScope(ArchiveScope::class)->where($this->getArchivablePageStateAttribute(), PageState::archived);
+        $query->withoutGlobalScope(ArchiveScope::class)->where(static::getArchiveColumn(), static::getArchivedValue());
     }
 
     public function scopeUnarchived($query)
     {
-        $query->withoutGlobalScope(ArchiveScope::class)->where($this->getArchivablePageStateAttribute(), '<>', PageState::archived);
+        $query->withoutGlobalScope(ArchiveScope::class)->where(static::getArchiveColumn(), '<>', static::getArchivedValue());
     }
 
     public function scopeWithArchived($query)