+++ src/ManagedModels/States/Archivable/ArchiveScope.php
@@ -7,12 +7,22 @@ namespace Thinktomorrow\Chief\ManagedModels\States\Archivable;
 use Illuminate\Database\Eloquent\Builder;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Eloquent\Scope;
-use Thinktomorrow\Chief\ManagedModels\States\PageState\PageState;
 
 class ArchiveScope implements Scope
 {
+    private string $archiveColumn;
+
+    private string $archivedValue;
+
+    public function __construct(string $archiveColumn, string $archivedValue)
+    {
+        // Constructor can be expanded in the future if needed
+        $this->archiveColumn = $archiveColumn;
+        $this->archivedValue = $archivedValue;
+    }
+
     public function apply(Builder $builder, Model $model)
     {
-        $builder->where($model->getTable().'.'.PageState::KEY, '<>', PageState::archived);
+        $builder->where($model->getTable().'.'.$this->archiveColumn, '<>', $this->archivedValue);
     }
 }